+++ admin/places.php
@@ -78,10 +78,13 @@ adminHeader('Zajímavá místa');
     <?php endif; ?>
   </p>
 <?php else: ?>
+  <?= bulkFormOpen('places', 'places.php') ?>
+  <?= bulkActionBar() ?>
   <table>
     <caption>Přehled zajímavých míst</caption>
     <thead>
       <tr>
+        <th scope="col"><input type="checkbox" class="bulk-select-all" aria-label="Vybrat vše"></th>
         <th scope="col">Místo</th>
         <th scope="col">Typ a lokalita</th>
         <th scope="col">Stav</th>
@@ -91,6 +94,7 @@ adminHeader('Zajímavá místa');
     <tbody>
     <?php foreach ($places as $place): ?>
       <tr>
+        <td><input type="checkbox" name="ids[]" value="<?= (int)$place['id'] ?>" class="bulk-checkbox" aria-label="Vybrat <?= h((string)$place['name']) ?>"></td>
         <td>
           <strong><?= h((string)$place['name']) ?></strong><br>
           <small style="color:#555">/places/<?= h((string)$place['slug']) ?></small>
@@ -144,8 +148,8 @@ adminHeader('Zajímavá místa');
     <?php endforeach; ?>
     </tbody>
   </table>
+  <?= bulkFormClose() ?>
+  <?= bulkCheckboxJs() ?>
 <?php endif; ?>
-
-
 
 <?php adminFooter(); ?>