+++ admin/polls.php
@@ -76,10 +76,13 @@ adminHeader('Ankety');
     <?php endif; ?>
   </p>
 <?php else: ?>
+  <?= bulkFormOpen('polls', 'polls.php') ?>
+  <?= bulkActionBar(false) ?>
   <table>
     <caption>Přehled anket</caption>
     <thead>
       <tr>
+        <th scope="col"><input type="checkbox" class="bulk-select-all" aria-label="Vybrat vše"></th>
         <th scope="col">Otázka</th>
         <th scope="col">Stav</th>
         <th scope="col">Hlasy</th>
@@ -99,6 +102,7 @@ adminHeader('Ankety');
         };
       ?>
       <tr<?= $state === 'scheduled' ? ' style="background:#eef6ff"' : '' ?>>
+        <td><input type="checkbox" name="ids[]" value="<?= (int)$poll['id'] ?>" class="bulk-checkbox" aria-label="Vybrat <?= h((string)$poll['question']) ?>"></td>
         <td>
           <strong><?= h((string)$poll['question']) ?></strong><br>
           <small style="color:#555">/polls/<?= h((string)$poll['slug']) ?></small>
@@ -126,8 +130,8 @@ adminHeader('Ankety');
     <?php endforeach; ?>
     </tbody>
   </table>
+  <?= bulkFormClose() ?>
+  <?= bulkCheckboxJs() ?>
 <?php endif; ?>
-
-
 
 <?php adminFooter(); ?>